+++ backend-core/src/main/java/com/promptpicture/backend/core/delle/adapter/DellEAdapter.java
@@ -1,6 +1,10 @@
 package com.promptpicture.backend.core.delle.adapter;
 
+import com.promptpicture.backend.core.prompt.domain.CreateIndividualPrompt;
+
 public interface DellEAdapter {
 
     String generatePicture(String promptText);
+
+    String generateIndividualPrompt(CreateIndividualPrompt createIndividualPrompt);
 }