+++ backend-core/src/main/java/com/promptpicture/backend/core/prompt/PromptFacade.java
@@ -1,5 +1,7 @@
 package com.promptpicture.backend.core.prompt;
 
+import com.promptpicture.backend.core.prompt.domain.CreateIndividualPrompt;
+import com.promptpicture.backend.core.prompt.domain.IndividualPrompt;
 import com.promptpicture.backend.core.prompt.domain.Prompt;
 import com.promptpicture.backend.core.prompt.domain.PromptFilter;
 import com.promptpicture.backend.core.prompt.use_case.CreatePromptUseCase;
@@ -7,6 +9,7 @@ import com.promptpicture.backend.core.prompt.use_case.GeneratePictureUseCase;
 import com.promptpicture.backend.core.prompt.use_case.GetListOfPromptByUserIdUseCase;
 import com.promptpicture.backend.core.prompt.use_case.GetListOfPromptUseCase;
 import com.promptpicture.backend.core.prompt.use_case.GetPromptDetailUseCase;
+import com.promptpicture.backend.core.prompt.use_case.SaveIndividualPromptUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -22,9 +25,14 @@ public class PromptFacade {
     private final GetPromptDetailUseCase getPromptDetailUseCase;
     private final CreatePromptUseCase createPromptUseCase;
     private final GetListOfPromptByUserIdUseCase getListOfPromptByUserIdUseCase;
+    private final SaveIndividualPromptUseCase saveIndividualPromptUseCase;
 
     public Prompt create(String promptText, UUID userId) {
        return generatePictureUseCase.execute(promptText, userId);
+    }
+
+    public IndividualPrompt saveIndividualPrompt(CreateIndividualPrompt createIndividualPrompt) {
+        return saveIndividualPromptUseCase.execute(createIndividualPrompt);
     }
 
     public List<Prompt> getListOfPromptUseCase(PromptFilter promptFilter) {