+++ backend-core/src/main/java/com/promptpicture/backend/core/prompt/domain/CreateIndividualPrompt.java
@@ -1,0 +1,16 @@
+package com.promptpicture.backend.core.prompt.domain;
+
+import lombok.Builder;
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+@Builder
+public class CreateIndividualPrompt {
+
+    private UUID externalCustomerId;
+    private String resolution;
+    private String promptText;
+
+}