+++ entrypoints/prompt-rest-api/src/main/java/com/promptpicture/backend/entrypoint/rest/model/input/CreateIndividualPromptRequest.java
@@ -1,0 +1,14 @@
+package com.promptpicture.backend.entrypoint.rest.model.input;
+
+import lombok.Data;
+
+import java.util.UUID;
+
+@Data
+public class CreateIndividualPromptRequest {
+
+    private UUID userId;
+    private String resolution;
+    private String promptText;
+
+}